2.5.2.0

+++ EditDistributedLoadDlg.cpp
@@ -256,19 +256,19 @@ BOOL CEditDistributedLoadDlg::OnInitDialog()
 
     if (m_Load.m_Span==ALL_SPANS)
     {
-       m_SpanCB.SetCurSel(nSpans);
+       m_SpanCB.SetCurSel((int)nSpans);
     }
     else
     {
       if ( 0 <= m_Load.m_Span && m_Load.m_Span < nSpans)
       {
-         m_SpanCB.SetCurSel(m_Load.m_Span);
+         m_SpanCB.SetCurSel((int)m_Load.m_Span);
       }
       else
       {
          ::AfxMessageBox(_T("Warning - The Span for this load is out of range. Resetting to Span 1"));
          m_Load.m_Span = 0;
-         m_SpanCB.SetCurSel(m_Load.m_Span);
+         m_SpanCB.SetCurSel((int)m_Load.m_Span);
       }
     }
 
@@ -282,13 +282,13 @@ BOOL CEditDistributedLoadDlg::OnInitDialog()
     {
       if (0 <= m_Load.m_Girder && m_Load.m_Girder < GirderIndexType(m_GirderCB.GetCount()-1) )
       {
-         m_GirderCB.SetCurSel(m_Load.m_Girder);
+         m_GirderCB.SetCurSel((int)m_Load.m_Girder);
       }
       else
       {
          ::AfxMessageBox(_T("Warning - The Girder for this load is out of range. Resetting to girder A"));
          m_Load.m_Girder=0;
-         m_GirderCB.SetCurSel(m_Load.m_Girder);
+         m_GirderCB.SetCurSel((int)m_Load.m_Girder);
       }
     }
